Add sanity checking for perspective regions
If a region from a preceding frame is available, the region for the new frame is checked against the preceding region.  If the new region is found to differ too much, then preceding region is re-used for new frame.
This is found to deal effectively with the problematic frames in the project video.

Project writeup modified accordingly.